+++ app/processing/components/index.js
@@ -1,0 +1,13 @@
+const getAddress = require('./get-address')
+const getDetails = require('./get-details')
+const getDetailedFunding = require('./get-detailed-funding')
+const getScheme = require('./get-scheme')
+const getDetailedPayments = require('./get-detailed-payments')
+
+module.exports = {
+  getAddress,
+  getDetailedFunding,
+  getDetails,
+  getDetailedPayments,
+  getScheme
+}
