--- app/processing/statement/components/get-scheme.js
@@ -1,11 +1,0 @@
-const getScheme = (year, frequency, agreementNumber) => {
-  return {
-    name: 'Sustainable Farming Incentive',
-    shortName: 'SFI',
-    year: String(year),
-    frequency,
-    agreementNumber
-  }
-}
-
-module.exports = getScheme
